# Speedup `Uint::shr`