--- a/benches/uint.rs
+++ b/benches/uint.rs
@@ -529,14 +529,14 @@ fn bench_sqrt(c: &mut Criterion) {
 
 criterion_group!(
     benches,
-    // bench_random,
-    // bench_mul,
-    // bench_division,
-    // bench_gcd,
-    // bench_shl,
+    bench_random,
+    bench_mul,
+    bench_division,
+    bench_gcd,
+    bench_shl,
     bench_shr,
-    // bench_inv_mod,
-    // bench_sqrt
+    bench_inv_mod,
+    bench_sqrt
 );
 
 criterion_main!(benches);