--- a/src/limb/shr.rs
+++ b/src/limb/shr.rs
@@ -11,12 +11,6 @@ impl Limb {
         Limb(self.0 >> shift)
     }
 
-    /// Computes `self >> 1` and return the result and the carry (0 or `1 << HI_BIT`).
-    #[inline(always)]
-    pub(crate) const fn shr1(self) -> (Self, Self) {
-        (Self(self.0 >> 1), Self(self.0 << Self::HI_BIT))
-    }
-
     /// Computes `self >> shift` and returns the result as well as the carry: the `shift` _most_
     /// significant bits of the `carry` are equal to the `shift` _least_ significant bits of `self`.
     ///