@erik-3milabs 2025-01-24
@tarcieri, can you think of ways to better integrate the (helper) functions I wrote? This might improve the quality of the code base and extend the performance speed up to other `shr` operations.

Also, once you're more/less on board with this PR, I'd happy to code up the same improvement for `shl`.