@erik-3milabs 2025-01-24
Also, with this implementation in place, it becomes painfully apparent that `Limb::shr` is _not_ constant time, unlike what its name suggests.