@erik-3milabs 2025-01-24
> > Also, with this implementation in place, it becomes painfully apparent that Limb::shr is not constant time, unlike what its name suggests.
> 
> Could you elaborate? Does the Rust compiler introduce some checks for `>>` on primitives?

My apologies; perhaps I was too quick to jump to a conclusion here. Taking a step back, a better statement is: 

_Either `Limb::shr` is vartime, or it is possible to speed up `Uint::shr` even further._

## Why I thought `Limb::shr` is vartime...

Allow me to elaborate. I jumped to my earlier conclusion because of how `Uint::shr` has been implemented until now. The "old" implementation makes successive calls to `Uint::shr_vartime(1<<i)` for `i=0, 1, 2, ....` and then `Uint::select`s the result if the `i`th bit was set in `shift`. The fact that `Uint::shr` leverages it as a subroute strongly suggests that `Uint::shr_vartime` is a very non-constant time algorithm, and that not a single part of it is constant time.

Moreover, the idea to
```rs
fn shr(&self, shift: u32) -> Self {
     let (bit_shift, limb_shift) = (shift % Limb::BITS, shift / Limb::BITS);
     self.intra_limb_shr(bit_shift).limb_shr(limb_shift)
}
```
was so apparent to me that I assumed there must be a good reason this has not been implemented :see_no_evil: After all, first ideas are usually terrible ideas, especially when working on a large repo like this with much more intelligent brains than mine working on them.

## ... and how we might be able to speed up `Uint::shr` even further.

I did some benchmarking of my own (I'm not familiar enough with compilers such as Godbolt to use that stuff) and got very stable results when using `Limb::shr`. This confirms that it might indeed be constant time.

When we use it to replace the `Limb::carrying_shr` call in `Uint::intra_limb_carrying_shr`, we save another 44ns when benchmarking on my machine, dropping `U2048::shr` to ~70ns. Not bad if you ask me! :smile: